% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plainview-package.R
\docType{package}
\name{plainview-package}
\alias{plainview-package}
\title{Plot Raster Images Interactively on a Plain HTML Canvas}
\description{
Plot Raster Images Interactively on a Plain HTML Canvas
}
\details{
Provides methods for plotting potentially large (raster) images
interactively on a plain HTML canvas.
Supports plotting of RasterLayer, RasterStack, RasterBrick
(from package raster) as well as png files located on disk. Interactivity
includes zooming, panning, and mouse location information. In case of
multi-layer RasterStacks or RasterBricks, RBG image plots are created
(similar to raster::plotRGB - but interactive).
}
\author{
Tim Appelhans, Stephan Woellauer
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}
